+++ R/feature_groups-comparison.R
@@ -73,8 +73,9 @@ setMethod("$", "featureGroupsComparison", function(x, name)
 #' output and generate a consensus.
 #'
 #' @param \dots For \code{comparison}: \code{featureGroups} objects that should
-#'   be compared. Objects are automatically labelled by their
-#'   \code{\link{algorithm}}.
+#'   be compared. If the arguments are named (\emph{e.g.} \code{myGroups =
+#'   fGroups}) then these are used for labelling, otherwise objects are
+#'   automatically labelled by their \code{\link{algorithm}}.
 #'
 #'   For \code{plot}, \code{plotVenn}, \code{plotChord}: further options passed
 #'   to \code{plot}, \pkg{\link{VennDiagram}} plotting functions (\emph{e.g.}
@@ -191,7 +192,8 @@ setMethod("comparison", "featureGroups", function(..., groupAlgo, groupArgs = li
                           unique = TRUE, .var.name = "...", add = ac)
     checkmate::reportAssertions(ac)
 
-    names(fGroupsList) <- make.unique(sapply(fGroupsList, algorithm))
+    n <- getArgNames(..., def = sapply(fGroupsList, algorithm))
+    names(fGroupsList) <- make.unique(n)
 
     # convert feature groups to features
     featsFromGroups <- convertFeatureGroupsToFeatures(fGroupsList)