+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## September 2019
 
+* Labels for objects in a `featureGroupsComparison` can be labelled (useful for e.g. plotting)
 * Caching and progress bar support for suspect screening
 * Updated/Fixed JDK installation for installation script
 * Fixed missing pipe operator import (`%>%`)